requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.0
idna==3.10
python-decouple==3.8
requests==2.32.3
Stockstir==2.1.6
urllib3==2.2.3
